テスト時にloadしないよう修正
テスト時のデータ読み込み・保存に起因するエラーが発生するので、テスト環境では動作しないようにしました。

+++ memo.rb
@@ -63,7 +63,7 @@ class Memo
     end
 
     def load
-      return if File.empty?(JSON_FILE) || !File.exist?(JSON_FILE)
+      return if File.empty?(JSON_FILE) || !File.exist?(JSON_FILE) || ENV['APP_ENV'] == 'test'
 
       memos_json = {}
       File.open(JSON_FILE, 'r') { |file| memos_json = JSON.parse(file.readline, symbolize_names: true) }
